+++ src/transports/function_transport.cpp
@@ -1,4 +1,4 @@
-#include "function.hpp"
+#include "function_transport.hpp"
 
 using namespace sentry;
 using namespace transports;
