+++ src/transports/libcurl_transport.hpp
@@ -6,7 +6,7 @@
 #include <curl/easy.h>
 #include <chrono>
 #include "../worker.hpp"
-#include "base.hpp"
+#include "base_transport.hpp"
 
 namespace sentry {
 namespace transports {
