+++ src/transports/winhttp_transport.hpp
@@ -2,10 +2,10 @@
 #define SENTRY_TRANSPORTS_WINHTTP_HPP_INCLUDED
 #ifdef SENTRY_WITH_WINHTTP_TRANSPORT
 
-#include <Windows.h>
 #include <WinHttp.h>
+#include <Windows.h>
 #include "../worker.hpp"
-#include "base.hpp"
+#include "base_transport.hpp"
 
 namespace sentry {
 namespace transports {
