+++ src/unwind.hpp
@@ -10,7 +10,7 @@
 #endif
 
 #include <functional>
-#include "../internal.hpp"
+#include "internal.hpp"
 
 namespace sentry {
 namespace unwinders {
