+++ src/unwinders/backtrace_unwinder.cpp
@@ -1,4 +1,4 @@
-#include "base.hpp"
+#include "../unwind.hpp"
 #ifdef SENTRY_WITH_BACKTRACE_UNWINDER
 #include <execinfo.h>
 
@@ -6,9 +6,9 @@ using namespace sentry;
 
 size_t unwinders::unwind_stack(void *addr, void **ptrs, size_t max_frames) {
     if (addr) {
-        #ifdef MAC_OS_X_VERSION_10_14
+#ifdef MAC_OS_X_VERSION_10_14
         return backtrace_from_fp(addr, ptrs, max_frames);
-        #endif
+#endif
         return 0;
     } else {
         return backtrace(ptrs, max_frames);
